server/app/animal_breeders/api/v1/serializers/profile_serializers.py
from rest_framework import serializers
from django.contrib.auth import get_user_model

User = get_user_model()

class ProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['id', 'name', 'email']  # Отображаем нужные поля профиля пользователя
        read_only_fields = ['email']  # Предполагаем, что email нельзя изменить

    def update(self, instance, validated_data):
        instance.name = validated_data.get('name', instance.name)
        instance.save()
        return instance