server/requirements.txt
psycopg2-binary==2.8.5
django-cors-headers==3.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
Django==3.1.3
urllib3==1.25.10
w3lib==2.1.2
python-dotenv
django-redis